git-svn-id: svn://192.168.1.10/public/CoffeeMud@8798 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/Clans.java
@@ -314,7 +314,7 @@ public class Clans extends StdLibrary implements ClanManager
 	                	for(int i=0;i<votingRoles.size();i++)
 	                	{
 	                		Integer roleID=votingRoles.get(i);
-	                		roleNames[i]=C.getRoleName(roleID, true, true);
+	                		roleNames[i]=C.getRoleName(roleID.intValue(), true, true);
 	                	}
 	                	String list = CMLib.english().toEnglishStringList(roleNames);
                         clanAnnounce(mob,"The "+C.typeName()+" "+C.clanID()+" has a new matter to vote upon. "
